# openapi/egdata.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "egdata.app API",
    "version": "0.1.0-alpha.0",
    "description": "REST API for Epic Games Store catalog, pricing, search, and public egdata.app data. This contract starts with public stable REST endpoints and will expand as legacy routes are classified.",
    "contact": {
      "name": "egdata.app",
      "url": "https://egdata.app"
    }
  },
  "servers": [
    {
      "url": "https://api.egdata.app",
      "description": "Production"
    },
    {
      "url": "https://api-gcp.egdata.app",
      "description": "Production GCP mirror"
    },
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/offers": {
      "get": {
        "operationId": "listOffers",
        "tags": [
          "Offers"
        ],
        "summary": "List offers with regional price data",
        "parameters": [
          {
            "$ref": "#/components/parameters/country"
          },
          {
            "$ref": "#/components/parameters/locale"
          },
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ],
        "x-egdata-visibility": "public",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/items": {
      "get": {
        "operationId": "listItems",
        "tags": [
          "Items"
        ],
        "summary": "List catalog items",
        "description": "Returns paginated Epic catalog items sorted by last modification date.",
        "parameters": [
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ],
        "x-egdata-visibility": "public",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/free-games/search": {
      "get": {
        "operationId": "searchFreeGames",
        "tags": [
          "Free Games"
        ],
        "summary": "Search historical free game promotions",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "description": "Search text.",
            "schema": {
              "type": "string",
              "example": "fallout"
            }
          },
          {
            "$ref": "#/components/parameters/locale"
          },
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ],
        "x-egdata-visibility": "public",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/search/{id}": {
      "get": {
        "operationId": "getSavedSearchQuery",
        "tags": [
          "Search"
        ],
        "summary": "Get a saved legacy search query by hash",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Saved query hash returned by POST /search.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-egdata-visibility": "public",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchBody"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/sellers": {
      "get": {
        "operationId": "listSellers",
        "tags": [
          "Sellers"
        ],
        "summary": "List sellers",
        "x-egdata-visibility": "public",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SellerSummary"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "EpicSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "egdata.app browser session backed by Epic authentication."
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Endpoint-specific API key."
      }
    },
    "parameters": {
      "country": {
        "name": "country",
        "in": "query",
        "required": false,
        "description": "ISO country code used to resolve an Epic Games Store pricing region. Defaults to the EGDATA_COUNTRY cookie, then US.",
        "schema": {
          "type": "string",
          "minLength": 2,
          "maxLength": 2,
          "example": "US"
        }
      },
      "locale": {
        "name": "locale",
        "in": "query",
        "required": false,
        "description": "Exact BCP-47-style locale used to overlay localized public offer text fields. Defaults to en-US.",
        "schema": {
          "type": "string",
          "example": "es-ES"
        }
      },
      "limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Maximum number of records to return. Most list endpoints cap this server-side.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "example": 10
        }
      },
      "page": {
        "name": "page",
        "in": "query",
        "required": false,
        "description": "One-based page number.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "Offer": {
        "type": "object",
        "description": "Public Epic Games Store offer DTO.",
        "additionalProperties": true,
        "properties": {
          "id": {
            "type": "string"
          },
          "namespace": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "longDescription": {
            "type": "string",
            "nullable": true
          },
          "offerType": {
            "type": "string",
            "nullable": true
          },
          "developerDisplayName": {
            "type": "string",
            "nullable": true
          },
          "publisherDisplayName": {
            "type": "string",
            "nullable": true
          },
          "seller": {
            "$ref": "#/components/schemas/SellerSummary"
          },
          "tags": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "offerMappings": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "productSlug": {
            "type": "string",
            "nullable": true
          },
          "urlSlug": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "keyImages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Image"
            }
          },
          "price": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Price"
              }
            ],
            "nullable": true
          },
          "releaseDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "locale": {
            "type": "string",
            "description": "Requested locale applied to this offer response.",
            "example": "es-ES"
          },
          "localeStatus": {
            "type": "string",
            "description": "Whether this offer is canonical en-US data, localized with an exact locale record, or fell back to canonical text.",
            "enum": [
              "canonical",
              "localized",
              "fallback"
            ]
          },
          "canonicalLocale": {
            "type": "string",
            "nullable": true,
            "description": "Canonical source locale, present on non-canonical responses.",
            "example": "en-US"
          },
          "localization": {
            "allOf": [
              {
                "$ref": "#/components/schemas/OfferLocalizationMetadata"
              }
            ],
            "nullable": true
          }
        }
      },
      "OfferLocalizationMetadata": {
        "type": "object",
        "description": "Metadata from the exact offer localization record used for this response.",
        "additionalProperties": true,
        "properties": {
          "source": {
            "type": "string",
            "nullable": true
          },
          "fetchedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "sourceUpdatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "OfferListResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "elements",
          "page",
          "limit",
          "total"
        ],
        "properties": {
          "elements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Offer"
            }
          },
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        }
      },
...
      "SearchBody": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string"
          },
          "offerType": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "customAttributes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "seller": {
            "type": "string"
          },
          "sortBy": {
            "type": "string",
            "enum": [
              "releaseDate",
              "lastModifiedDate",
              "effectiveDate",
              "creationDate",
              "viewableDate",
              "pcReleaseDate",
              "upcoming",
              "priceAsc",
              "priceDesc",
              "price",
              "discount",
              "discountPercent",
              "giveawayDate"
            ]
          },
          "sortDir": {
            "type": "string",
            "enum": [
              "asc",
              "desc"
            ]
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "refundType": {
            "type": "string"
          },
          "isCodeRedemptionOnly": {
            "type": "boolean"
          },
          "price": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "min": {
                "type": "number"
              },
              "max": {
                "type": "number"
              }
            }
          },
          "onSale": {
            "type": "boolean"
          },
          "developerDisplayName": {
            "type": "string"
          },
          "publisherDisplayName": {
            "type": "string"
          },
          "spt": {
            "type": "boolean"
          },
          "excludeBlockchain": {
            "type": "boolean"
          },
          "pastGiveaways": {
            "type": "boolean"
          },
          "isLowestPrice": {
            "type": "boolean"
          },
          "isLowestPriceEver": {
            "type": "boolean"
          }
        }
      },
      "Item": {
        "type": "object",
        "description": "Public Epic Games Store item DTO.",
        "additionalProperties": true,
        "properties": {
          "id": {
            "type": "string"
          },
          "namespace": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "keyImages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Image"
            }
          },
          "releaseInfo": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
      "ItemListResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "elements",
          "page",
          "limit",
          "total"
        ],
        "properties": {
          "elements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          },
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        }
      },
...
      "Price": {
        "type": "object",
        "description": "Regional offer price from the price engine.",
        "additionalProperties": true,
        "properties": {
          "offerId": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "price": {
            "type": "object",
            "additionalProperties": true,
            "properties": {
              "currencyCode": {
                "type": "string"
              },
              "originalPrice": {
                "type": "number"
              },
              "discountPrice": {
                "type": "number"
              },
              "discount": {
                "type": "number"
              }
            }
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SellerSummary": {
        "type": "object",
        "description": "Seller metadata.",
        "additionalProperties": true,
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "Image": {
        "type": "object",
        "description": "Epic image asset reference.",
        "additionalProperties": true,
        "properties": {
          "type": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "width": {
            "type": "integer"
          },
          "height": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}